+++ NEWS.md
@@ -3,7 +3,8 @@
 * `top_freq()` now correctly selects bottoms items using negative a number for *n*
 * `freq.default()` is now exported for use in other packages
 * All numeric calculation in the header of frequency tables now use the same algorithm as used by Minitab and SPSS (see 'Type 6' on `stats::quantile()`)
-
+* More robust results for `clean_character()`, it also keeps in-between spaces now
+* `clean_numeric()` now supports currency
 
 # 1.0.0