+++ NEWS.md
@@ -1,6 +1,7 @@
 # 1.0.0.9000
 
 * Added support for currency as a new class: `as.currency()` and `clean_currency()`. They also come with 'S3 methods' for `print`, `format`, `sum`, `min` and `max`.
+* Added `clean_POSIXct()` to clean date/time objects
 * `top_freq()` now correctly selects bottoms items using negative a number for *n*
 * `freq.default()` is now exported for use in other packages
 * All numeric calculation in the header of frequency tables now use the same algorithm as used by Minitab and SPSS (see 'Type 6' on `stats::quantile()`)